feat(#tdd): add a naive / failing AssignmentGenerator impl.
this commit should never have existed if not for teaching purposes.
this naive implementation simply fails to do anything in acceptable time
whenever the number of combinations becomes relatively high.
TDDwise, we now have an implementation and some tests to work with ; that
should be easy to refine the algorithm and cut complexity or time down