feat(#tdd): add a new constraint
avoid assigning the same task type to someone twice in a row.
+ use it in PlanningController
+ a new utility field in BacktrackableAssignment
+ some doc

this new constraint is pretty much useless to the current algorithm in
most cases, thanks to the choosePerson heuristics.
Nevertheless, it helps rejecting candidates when
the number of people = the number
of tasks
(41k backtrackings vs only 103 for a 4 types-4 people problem with 6 games,
and a much better solution)